Move VictoireWebTestCase into Tests/Functional folder
